+++ packages/react/src/components/AppShell/src/components/AppShellChat/index.ts
@@ -1,2 +1,5 @@
 export { AppShellChat } from "./AppShellChat";
-export type { AppShellChatProps } from "./AppShellChat.types";
+export type {
+  AppShellChatProperties,
+  AppShellChatProps,
+} from "./AppShellChat.types";